Moved X11Utils into UICommon.
