+++ Source/Core/DolphinWX/Config/InterfaceConfigPane.cpp
@@ -30,7 +30,7 @@
 #include "DolphinWX/WxUtils.h"
 
 #if defined(HAVE_XRANDR) && HAVE_XRANDR
-#include "DolphinWX/X11Utils.h"
+#include "UICommon/X11Utils.h"
 #endif
 
 static const std::array<std::string, 29> language_ids{{
