+++ Source/Core/DolphinWX/MainNoGUI.cpp
@@ -168,7 +168,7 @@ void Host_YieldToUI()
 
 #if HAVE_X11
 #include <X11/keysym.h>
-#include "DolphinWX/X11Utils.h"
+#include "UICommon/X11Utils.h"
 
 class PlatformX11 : public Platform
 {
